requirements.txt
pandas
yahoo_fin
requests_html
numpy
dash==1.12.0
beautifulsoup4
boto3